Add information related to using Unicode
As from Erlang/OTP R16B you can select the ISO Latin-1 range or the whole Unicode range by supplying startup flag +pc latin1 or +pc unicode, respectively. For backward compatibility, latin1 is default. This only controls how heuristic string detection is done. More ranges are expected to be added in the future, enabling tailoring of the heuristics to the language and region relevant to the user.